Fix job_attachments migration: use BIGINT UNSIGNED for foreign keys
- Change attachment_id to BIGINT UNSIGNED to match users.userID
- Change uploaded_by to BIGINT UNSIGNED to match users.userID (bigint unsigned)
- Keep job_id as INT to match jobs.jobID (int)

+++ migrations/023_create_job_attachments_table.up.sql
@@ -1,13 +1,13 @@
 -- Create job_attachments table for file attachments to jobs
 CREATE TABLE job_attachments (
-    attachment_id INT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
+    attachment_id BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
     job_id INT NOT NULL,
     filename VARCHAR(255) NOT NULL,
     original_filename VARCHAR(255) NOT NULL,
     file_path VARCHAR(500) NOT NULL,
     file_size BIGINT NOT NULL,
     mime_type VARCHAR(100) NOT NULL,
-    uploaded_by INT UNSIGNED DEFAULT NULL,
+    uploaded_by BIGINT UNSIGNED DEFAULT NULL,
     uploaded_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     description TEXT,
     is_active BOOLEAN DEFAULT TRUE,
